refactor: move chat realtime ownership
