+++ messaging/realtime/__init__.py
@@ -1,0 +1,1 @@
+"""Realtime chat support owned by the messaging module."""
